+++ AllProfessions/docs/release-notes.md
@@ -3,6 +3,7 @@
 # Release notes
 ## Upcoming release
 * Added `config.json` file to disable professions.
+* Added in-game config UI via [Generic Mod Config Menu](https://www.nexusmods.com/stardewvalley/mods/5098).
 
 ## 1.4.4
 Released 26 March 2021 for SMAPI 3.9.5 or later. Updated by Pathoschild.